+++ tests/Feature/Form/FormField/FormFieldTest.php
@@ -2,6 +2,7 @@
 
 namespace Portavice\Bladestrap\Tests\Feature\Form\FormField;
 
+use Illuminate\Support\Facades\Config;
 use PHPUnit\Framework\Attributes\DataProvider;
 use Portavice\Bladestrap\Tests\Feature\ComponentTestCase;
 use Portavice\Bladestrap\Tests\Traits\TestsBooleanAttributes;
@@ -65,14 +66,41 @@ class FormFieldTest extends ComponentTestCase
     }
 
     #[DataProvider('booleanFormFieldAttributes')]
-    public function testFormFieldWithBooleanAttributesRendersCorrectly(string $html, string $blade): void
+    public function testFormFieldWithBooleanAttributesRendersCorrectly(string $html, string $blade, string $additionalLabel): void
     {
         $this->assertBladeRendersToHtml(
             '<div class="mb-3">
-                <label for="first_name" class="form-label">First name</label>
+                <label for="first_name" class="form-label">First name' . $additionalLabel . '</label>
                 <input id="first_name" name="first_name" type="text" value="Patrick" class="form-control" '. $html . '/>
             </div>',
             '<x-bs::form.field name="first_name" type="text" value="Patrick" ' . $blade . '>First name</x-bs::form.field>'
         );
+    }
+
+    #[DataProvider('conditionsForRequiredMarker')]
+    public function testRequiredFormFieldOnlyMarkedAsRequiredIfEnabled(string $html, string $blade): void
+    {
+        Config::set('bladestrap.form.field.mark_as_required', false);
+        $this->assertBladeRendersToHtml(
+            '<div class="mb-3">
+                <label for="first_name" class="form-label">First name' . $html . '</label>
+                <input id="first_name" name="first_name" type="text" value="Patrick" class="form-control" required/>
+            </div>',
+            $blade
+        );
+    }
+
+    public static function conditionsForRequiredMarker(): array
+    {
+        return [
+            [
+                '',
+                '<x-bs::form.field name="first_name" type="text" value="Patrick" :required="true">First name</x-bs::form.field>',
+            ],
+            [
+                ' *',
+                '<x-bs::form.field name="first_name" type="text" value="Patrick" :required="true" :mark-as-required="true">First name</x-bs::form.field>',
+            ],
+        ];
     }
 }