+++ tests/Feature/Form/FormField/Type/RadioTest.php
@@ -90,11 +90,11 @@ class RadioTest extends ComponentTestCase
     }
 
     #[DataProvider('booleanFormFieldAttributes')]
-    public function testRadioWithBooleanAttributesRendersCorrectly(string $html, string $blade): void
+    public function testRadioWithBooleanAttributesRendersCorrectly(string $html, string $blade, string $additionalLabel): void
     {
         $this->assertBladeRendersToHtml(
             '<div class="mb-3">
-                <label for="my_model" class="form-label">My Model</label>
+                <label for="my_model" class="form-label">My Model' . $additionalLabel . '</label>
                 <div class="form-check">
                     <input id="my_model-1" name="my_model" type="radio" value="1" class="form-check-input" ' . $html . '/>
                     <label class="form-check-label" for="my_model-1">A</label>